pyproject.toml
[project]
name = "free-agent"
version = "0.1.0"
description = "Terminal-first base for building local agent and multi-agent systems on LangChain deepagents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Robert Gomez" }]
keywords = ["llm", "agents", "deepagents", "langchain", "ollama", "cli", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Terminals",
]

dependencies = [
    "deepagents>=0.0.20",
    "langchain-anthropic>=0.3.0",
    "langchain-ollama>=0.2.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "prompt-toolkit>=3.0.47",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/Robert-Gomez-AI/free-agent"
Repository = "https://github.com/Robert-Gomez-AI/free-agent"
Issues = "https://github.com/Robert-Gomez-AI/free-agent/issues"

[project.scripts]
free-agent = "free_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/free_agent"]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
